pyproject.toml
[tool.poetry]
name = "aoe-29-discordbot"
version = "0.1.0"
description = ""
authors = ["noracami <kerke.2011@gmail.com>"]
readme = "README.md"
packages = [{include = "aoe_29_discordbot"}]

[tool.poetry.dependencies]
python = "^3.11"
discord-py-interactions = "^5.11.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
